+++ src/evm-lite-core/src/classes/accounts/Accounts.ts
@@ -1,15 +1,14 @@
 import { Accounts as Web3Accounts } from 'web3-eth-accounts';
 
-import { Address, AddressType, Gas, GasPrice, Value } from '../../types';
-import Transaction, { BaseTransaction } from '../transaction/Transaction';
 import { V3JSONKeyStore } from './Account';
 
-import AccountClient from '../../clients/AccountClient';
-import Account from './Account';
+import { Defaults } from '../EVMLC';
 
-interface AccountDefaultOptions extends BaseTransaction {
-	from: Address;
-}
+import Transaction, { BaseTransaction } from '../transaction/Transaction';
+
+import AccountClient from '../../clients/AccountClient';
+import EVM from '../../types';
+import Account from './Account';
 
 export default class Accounts extends AccountClient {
 	private accounts: Web3Accounts;
@@ -19,13 +18,9 @@ export default class Accounts extends AccountClient {
 	 *
 	 * @param host - The host of the active node.
 	 * @param port - The port of the HTTP service.
-	 * @param contractOptions - The default options for accounts
+	 * @param defaults - The default options for accounts
 	 */
-	constructor(
-		host: string,
-		port: number,
-		private accountOptions: AccountDefaultOptions
-	) {
+	constructor(host: string, port: number, public defaults: Defaults) {
 		super(host, port);
 
 		this.accounts = new Web3Accounts('http://', {
@@ -43,9 +38,7 @@ export default class Accounts extends AccountClient {
 	 * `signTransaction()` attribute. Allowing to sign transactions.
 	 *
 	 * ```typescript
-	 * // Create and encrypt a new account with a password
 	 * const keystore = evmlc.accounts.create().encrypt('password');
-	 * // Decrypt the account with corresponding password
 	 * const decrypted = evmlc.accounts.decrypt(keystore, 'password');
 	 * ```
 	 *
@@ -64,7 +57,6 @@ export default class Accounts extends AccountClient {
 	 *
 	 * @remarks
 	 * ```typescript
-	 * // Create a new account
 	 * const account = evmlc.accounts.create();
 	 * ```
 	 *
@@ -82,15 +74,13 @@ export default class Accounts extends AccountClient {
 	 * address.
 	 *
 	 * @remarks
-	 * This function will also fetch the nonce from the node with connection
-	 * details specified in the contructor for this class.
+	 * This function will not fetch nonce from the node. The example shows
+	 * how to make a trnasfer of 200 tokens.
 	 *
 	 * ```typescript
 	 * const transfer = async () {
-	 *     // Prepare a transfer transaction to submitted after signing
-	 *     const transaction = await evmlc.prepareTransfer('TO_ADDRESS', 200);
-	 *     // Sign the transaction with a new  account and submit to node.
-	 *     await transaction.submit({}, evmlc.accounts.create())
+	 *     const transaction = evmlc.prepareTransfer('TO_ADDRESS', 200);
+	 *     await transaction.submit(evmlc.accounts.create())
 	 * }
 	 * ```
 	 *
@@ -99,15 +89,13 @@ export default class Accounts extends AccountClient {
 	 * @param from - Overrides `from` address set in the constructor.
 	 */
 	public prepareTransfer(
-		to: string,
-		value: Value,
-		from?: string
-	): Promise<Transaction> {
-		const fromObject = new AddressType(
-			(from || this.accountOptions.from.value).trim()
-		);
+		to: EVM.Address,
+		value: EVM.Value,
+		from?: EVM.Address
+	): Transaction {
+		const _from = (from || this.defaults.from).trim();
 
-		if (!fromObject.value) {
+		if (!_from) {
 			throw new Error(
 				'Default `from` address cannot be left blank or empty.'
 			);
@@ -123,81 +111,17 @@ export default class Accounts extends AccountClient {
 			);
 		}
 
-		return this.getAccount(fromObject.value).then(
-			account =>
-				new Transaction(
-					{
-						from: fromObject,
-						to: new AddressType(to.trim()),
-						value,
-						gas: this.accountOptions.gas,
-						gasPrice: this.accountOptions.gasPrice,
-						nonce: account.nonce,
-						chainId: 1
-					},
-					this.host,
-					this.port,
-					false
-				)
+		return new Transaction(
+			{
+				from: _from,
+				to: to.trim(),
+				value,
+				gas: this.defaults.gas,
+				gasPrice: this.defaults.gasPrice
+			},
+			this.host,
+			this.port,
+			false
 		);
-	}
-
-	/**
-	 * The defaults for contracts created from this object.
-	 */
-	get defaults() {
-		return {
-			from: this.accountOptions.from.value,
-			gas: this.accountOptions.gas,
-			gasPrice: this.accountOptions.gasPrice
-		};
-	}
-
-	/**
-	 * The default `from` address that will be used for any transactions
-	 * created from this object.
-	 */
-	get defaultFrom(): string {
-		return this.accountOptions.from.value;
-	}
-
-	/**
-	 * Set the default `from` to be used for any transactions created from
-	 * this object.
-	 */
-	set defaultFrom(address: string) {
-		this.accountOptions.from = new AddressType(address);
-	}
-
-	/**
-	 * The default `gas` value that will be used for any transactions created
-	 * from this object.
-	 */
-	get defaultGas(): Gas {
-		return this.accountOptions.gas;
-	}
-
-	/**
-	 * Set the default `gas` value to be used for any transactions created from
-	 * this object.
-	 */
-	set defaultGas(gas: Gas) {
-		this.accountOptions.gas = gas;
-	}
-
-	/**
-	 * The default `gasPrice` value that will be used for any transactions
-	 * created from this object.
-	 */
-	get defaultGasPrice(): GasPrice {
-		return this.accountOptions.gasPrice;
-	}
-
-	/**
-	 * Set the default `gasPrice` to be used for any transactions created from
-	 * this object.
-	 */
-	set defaultGasPrice(gasPrice: GasPrice) {
-		this.accountOptions.gasPrice = gasPrice;
 	}
 }