+++ src/evm-lite-core/tests/classes/Contracts.test.ts
@@ -1,3 +1,5 @@
+import Defaults from '../vars';
+
 import { EVMLC } from '../../src';
 
 let contract: any;
@@ -5,54 +7,24 @@ let evmlc: EVMLC;
 
 describe('Contracts.ts', () => {
 	beforeEach(() => {
-		evmlc = new EVMLC('127.0.0.1', 8080, {
-			from: '0X5E54B1907162D64F9C4C7A46E3547084023DA2A0',
-			gas: 10000000,
-			gasPrice: 0
+		evmlc = new EVMLC(Defaults.HOST, Defaults.POST, {
+			from: Defaults.FROM,
+			gas: Defaults.GAS,
+			gasPrice: Defaults.GAS_PRICE
 		});
 	});
 
-	it('should have the default gas from `EVMLC`', () => {
-		expect(evmlc.contracts.defaults.gas).toBe(evmlc.defaultGas);
-	});
-
-	it('should have the default gasPrice from `EVMLC`', () => {
-		expect(evmlc.contracts.defaults.gasPrice).toBe(evmlc.defaultGasPrice);
-	});
-
-	it('should have the default `from` from `EVMLC`', () => {
-		expect(evmlc.contracts.defaults.from).toBe(evmlc.defaultFrom);
-	});
-
-	it('should have the default gas from `EVMLC` after change', () => {
-		evmlc.defaultGas = 9999;
-
-		expect(evmlc.contracts.defaults.gas).toBe(9999);
-	});
-
-	it('should have the default gasPrice from `EVMLC` after change', () => {
-		evmlc.defaultGasPrice = 12;
-
-		expect(evmlc.contracts.defaults.gasPrice).toBe(12);
-	});
-
-	it('should have the default `from` from `EVMLC` after change', () => {
-		evmlc.defaultFrom = 'ASD';
-
-		expect(evmlc.contracts.defaults.from).toBe('ASD');
-	});
-
 	it('contract should have default gas', async () => {
-		contract = await evmlc.contracts.load([]);
+		contract = evmlc.contracts.load([]);
 
-		expect(contract.options.gas).toBe(evmlc.defaultGas);
+		expect(contract.contractOptions.gas).toBe(evmlc.defaultGas);
 	});
 
 	it('contract should have default gasPrice', async () => {
-		expect(contract.options.gasPrice).toBe(evmlc.defaultGasPrice);
+		expect(contract.contractOptions.gasPrice).toBe(evmlc.defaultGasPrice);
 	});
 
 	it('contract should have default from', async () => {
-		expect(contract.options.from.value).toBe(evmlc.defaultFrom);
+		expect(contract.contractOptions.from).toBe(evmlc.defaultFrom);
 	});
 });