+++ src/evm-lite-lib/classes/Config.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as toml from 'toml';
 import * as tomlify from 'tomlify-j0.4';
 
-import { Gas, GasPrice } from 'evm-lite-core';
+import { EVM } from 'evm-lite-core';
 
 import Static from './Static';
 
@@ -13,9 +13,9 @@ export interface ConfigSchema {
 		port: number;
 	};
 	defaults: {
-		from: string;
-		gas: Gas;
-		gasPrice: GasPrice;
+		from: EVM.Address;
+		gas: EVM.Gas;
+		gasPrice: EVM.GasPrice;
 	};
 	storage: {
 		keystore: string;
@@ -23,16 +23,10 @@ export interface ConfigSchema {
 }
 
 export default class Config {
-	public readonly path: string;
-
 	public data: ConfigSchema;
 
-	constructor(
-		public readonly directory: string,
-		public readonly name: string
-	) {
+	constructor(public readonly path: string) {
 		this.data = this.default();
-		this.path = path.join(directory, name);
 
 		if (Static.exists(this.path)) {
 			const tomlData: string = fs.readFileSync(this.path, 'utf8');
@@ -48,6 +42,11 @@ export default class Config {
 	}
 
 	public default(): ConfigSchema {
+		const list = this.path.split('/');
+		list.pop();
+
+		const keystoreParent = list.join('/');
+
 		return {
 			connection: {
 				host: '127.0.0.1',
@@ -59,7 +58,7 @@ export default class Config {
 				gasPrice: 0
 			},
 			storage: {
-				keystore: path.join(this.directory, 'keystore')
+				keystore: path.join(keystoreParent, 'keystore')
 			}
 		};
 	}